add timepicker

+++ data/src/main/java/ru/vafeen/data/datastore/models/DataStoreSettings.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class DataStoreSettings(
-    val defaultExerciseDurationMillis: Int = 60000,
-    val defaultBreakDurationMillis: Int = 10000,
-    val exerciseDurationMillis: Int = 60000,
-    val breakDurationMillis: Int = 15000,
+    val defaultExerciseDurationMillis: Long = 60000,
+    val defaultBreakDurationMillis: Long = 10000,
+    val exerciseDurationMillis: Long = 60000,
+    val breakDurationMillis: Long = 15000,
 )
